+++ deno.jsonc
@@ -29,6 +29,7 @@
     }
   },
   "tasks": {
-    "udd": "deno run -A _tasks/udd.ts"
+    "udd": "deno run -A _tasks/udd.ts",
+    "moderate": "deno run -A https://deno.land/x/moderate@0.0.5/mod.ts && dprint fmt"
   }
 }
